+++ panda/src/gobj/texturePeeker.h
@@ -36,10 +36,10 @@ PUBLISHED:
   INLINE int get_y_size() const;
   INLINE int get_z_size() const;
 
-  INLINE bool has_pixel(size_t x, size_t y) const;
+  INLINE bool has_pixel(int x, int y) const;
   void lookup(LColor &color, PN_stdfloat u, PN_stdfloat v) const;
   void lookup(LColor &color, PN_stdfloat u, PN_stdfloat v, PN_stdfloat w) const;
-  void fetch_pixel(LColor &color, size_t x, size_t y) const;
+  void fetch_pixel(LColor &color, int x, int y) const;
   bool lookup_bilinear(LColor &color, PN_stdfloat u, PN_stdfloat v) const;
   void filter_rect(LColor &color,
                    PN_stdfloat min_u, PN_stdfloat min_v,
